requirements.txt
bleak==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
contourpy==1.3.3 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
dbus-fast==2.44.3 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Linux"
fonttools==4.59.2 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.9 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.10.6 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.3.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.3.2 ; python_version >= "3.12" and python_version < "4.0"
pillow==11.3.0 ; python_version >= "3.12" and python_version < "4.0"
pycairo==1.28.0 ; python_version >= "3.12" and python_version < "4.0"
pygobject==3.52.3 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
scipy==1.16.1 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
